migrate misskey action.

+++ core/src/commonMain/kotlin/work/socialhub/planetlink/model/request/PollForm.kt
@@ -8,7 +8,7 @@ class PollForm {
     var multiple: Boolean = false
 
     /** Expires in (min)  */
-    var expiresIn: Long = 1440L
+    var expiresIn: Int = 1440
 
     /**
      * Add poll option
